rename: kebab case 변경, nav height 수정
